# apps/backend/docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "LaVidaLuca Backend API",
    "description": "Backend API for LaVidaLuca - MFR training platform. This API provides endpoints for user authentication, activity management, contact form submissions, and AI-powered suggestions for collaborative learning experiences.",
    "version": "1.0.0",
    "contact": {
      "name": "LaVidaLuca Support",
      "email": "support@lavidaluca.com",
      "url": "https://lavidaluca.com/contact"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/contacts/": {
      "post": {
        "tags": ["Contacts"],
        "summary": "Submit contact form",
        "description": "Submit a new contact form (public endpoint)",
        "operationId": "create_contact",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactCreate"
              },
              "example": {
                "name": "Jean Dupont",
                "email": "jean.dupont@example.com",
                "phone": "+33123456789",
                "organization": "MFR de Provence",
                "subject": "Demande de partenariat",
                "message": "Bonjour, nous souhaitons explorer une collaboration avec La Vida Luca pour nos formations...",
                "contact_type": "partnership",
                "consent_privacy": true,
                "consent_marketing": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact request submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContactResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "Jean Dupont",
                    "email": "jean.dupont@example.com",
                    "subject": "Demande de partenariat",
                    "contact_type": "partnership",
                    "status": "new",
                    "priority": "normal",
                    "is_responded": false,
                    "created_at": "2024-01-15T10:30:00Z"
                  },
                  "message": "Contact request submitted successfully"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Contacts"],
        "summary": "List contact requests",
        "description": "List contact requests with filtering (admin only)",
        "operationId": "list_contacts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1,
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "contact_type",
            "in": "query",
            "required": false,
            "schema": {
              "title": "Contact Type",
              "enum": ["general", "partnership", "support", "feedback", "collaboration", "press"],
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "title": "Status",
              "enum": ["new", "in_progress", "resolved", "closed"],
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "description": "Search in name, email, subject, message"
          }
        ],
        "responses": {
          "200": {
            "description": "Contacts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePaginatedResponseContactListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/contacts/{contact_id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by ID",
        "description": "Get contact by ID (admin only)",
        "operationId": "get_contact",
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Contact Id",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContactResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Update contact",
        "description": "Update contact status and metadata (admin only)",
        "operationId": "update_contact",
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Contact Id",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactUpdate"
              },
              "example": {
                "status": "in_progress",
                "priority": "high",
                "tags": ["urgent", "partnership"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContactResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete contact",
        "description": "Delete contact (admin only)",
        "operationId": "delete_contact",
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Contact Id",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDict"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ContactCreate": {
        "title": "ContactCreate",
        "required": ["name", "email", "subject", "message"],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "maxLength": 255
          },
          "email": {
            "title": "Email",
            "type": "string",
            "format": "email"
          },
          "phone": {
            "title": "Phone",
            "type": "string",
            "maxLength": 50
          },
          "organization": {
            "title": "Organization",
            "type": "string",
            "maxLength": 255
          },
          "subject": {
            "title": "Subject",
            "type": "string",
            "maxLength": 500
          },
          "message": {
            "title": "Message",
            "type": "string",
            "minLength": 10
          },
          "contact_type": {
            "title": "Contact Type",
            "type": "string",
            "enum": ["general", "partnership", "support", "feedback", "collaboration", "press"],
            "default": "general"
          },
          "consent_privacy": {
            "title": "Consent Privacy",
            "type": "boolean",
            "default": true
          },
          "consent_marketing": {
            "title": "Consent Marketing",
            "type": "boolean",
            "default": false
          },
          "metadata": {
            "title": "Metadata",
            "type": "object",
            "default": {}
          }
        },
        "description": "Schema for creating a contact form submission"
      },
      "ContactUpdate": {
        "title": "ContactUpdate",
        "type": "object",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string",
            "enum": ["new", "in_progress", "resolved", "closed"]
          },
          "priority": {
            "title": "Priority",
            "type": "string",
            "enum": ["low", "normal", "high", "urgent"]
          },
          "assigned_to": {
            "title": "Assigned To",
            "type": "string",
            "format": "uuid"
          },
          "tags": {
            "title": "Tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          }
        },
        "description": "Schema for updating contact metadata (admin only)"
      },
      "ContactResponse": {
        "title": "ContactResponse",
        "required": ["id", "name", "email", "subject", "message", "contact_type", "status", "priority", "is_responded", "response_count", "metadata", "tags", "consent_privacy", "consent_marketing"],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string",
            "format": "email"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "organization": {
            "title": "Organization",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "contact_type": {
            "title": "Contact Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "assigned_to": {
            "title": "Assigned To",
            "type": "string",
            "format": "uuid"
          },
          "is_responded": {
            "title": "Is Responded",
            "type": "boolean"
          },
          "response_count": {
            "title": "Response Count",
            "type": "integer"
          },
          "last_response_at": {
            "title": "Last Response At",
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "tags": {
            "title": "Tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "consent_privacy": {
            "title": "Consent Privacy",
            "type": "boolean"
          },
          "consent_marketing": {
            "title": "Consent Marketing",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "Complete contact information response"
      },
      "ContactListResponse": {
        "title": "ContactListResponse",
        "required": ["id", "name", "email", "subject", "contact_type", "status", "priority", "is_responded"],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string",
            "format": "email"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "contact_type": {
            "title": "Contact Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "is_responded": {
            "title": "Is Responded",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "Simplified contact information for listing"
      },
      "ApiResponseContactResponse": {
        "title": "ApiResponse[ContactResponse]",
        "required": ["success", "data"],
        "type": "object",
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ContactResponse"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "ApiResponsePaginatedResponseContactListResponse": {
        "title": "ApiResponse[PaginatedResponse[ContactListResponse]]",
        "required": ["success", "data"],
        "type": "object",
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/PaginatedResponseContactListResponse"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "PaginatedResponseContactListResponse": {
        "title": "PaginatedResponse[ContactListResponse]",
        "required": ["items", "total", "page", "size", "pages"],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactListResponse"
            }
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          }
        }
      },
      "ApiResponseDict": {
        "title": "ApiResponse[Dict]",
        "required": ["success", "data"],
        "type": "object",
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "data": {
            "title": "Data",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login endpoint"
      }
    }
  },
...
}